python/requirements.txt
requests

# real-time
rel
websocket-client